+++ plugins/integrations/kubernetes-service/src/main/resources/conf/k8s-control-node.yml
@@ -271,21 +271,22 @@ write_files:
       echo "true" > /home/cloud/success
 
   - path: /opt/bin/setup-containerd
-      permissions: '0755'
-      owner: root:root
-      content: |
-        #!/bin/bash -e
+    permissions: '0755'
+    owner: root:root
+    content: |
+      #!/bin/bash -e
+      
+      export registryConfig="\\        [plugins.\"io.containerd.grpc.v1.cri\".registry.mirrors.\"{{registry.url.endpoint}}\"]\n \\         endpoint = [\"{{registry.url}}\"]"
+      export registryCredentials="\\      [plugins.\"io.containerd.grpc.v1.cri\".registry.configs.\"{{registry.url.endpoint}}\".auth]\n \tauth = \"\" \n \tidentitytoken = \"{{registry.token}}\""
+      
+      echo "creating config file for containerd"
+      containerd config default > /etc/containerd/config.toml
+      sed  -i '/\[plugins."io.containerd.grpc.v1.cri".registry\]/a '"${registryCredentials}"'' /etc/containerd/config.toml
+      sed  -i '/\[plugins."io.containerd.grpc.v1.cri".registry.mirrors\]/a '"${registryConfig}"'' /etc/containerd/config.toml
+      
+      echo "Restarting containerd service"
+      systemctl restart containerd
 
-        export registryConfig="\\        [plugins.\"io.containerd.grpc.v1.cri\".registry.mirrors.\"{{registry.url.endpoint}}\"]\n \\         endpoint = [\"{{registry.url}}\"]"
-        export registryCredentials="\\      [plugins.\"io.containerd.grpc.v1.cri\".registry.configs.\"{{registry.url.endpoint}}\".auth]\n \tauth = \"\" \n \tidentitytoken = \"{{registry.token}}\""
-
-        echo "creating config file for containerd"
-        containerd config default > /etc/containerd/config.toml
-        sed  -i '/\[plugins."io.containerd.grpc.v1.cri".registry\]/a '"${registryCredentials}"'' /etc/containerd/config.toml
-        sed  -i '/\[plugins."io.containerd.grpc.v1.cri".registry.mirrors\]/a '"${registryConfig}"'' /etc/containerd/config.toml
-
-        echo "Restarting containerd service"
-        systemctl restart containerd
 
   - path: /etc/systemd/system/setup-kube-system.service
     permissions: '0755'
