+++ tools/appliance/systemvmtemplate/scripts/configure_systemvm_services.sh
@@ -19,7 +19,7 @@
 set -e
 set -x
 
-CLOUDSTACK_RELEASE=4.16.0
+CLOUDSTACK_RELEASE=4.16.1
 
 function configure_apache2() {
    # Enable ssl, rewrite and auth
